updates to server.js and handlebars
