+++ package.json
@@ -17,6 +17,7 @@
     "express": "^4.17.1",
     "express-handlebars": "^8.0.1",
     "express-session": "^1.18.0",
+    "handlebars": "^4.7.8",
     "pg": "^8.11.3",
     "sequelize": "^6.33.0"
   }
