+++ server.js
@@ -1,6 +1,10 @@
 const express = require('express');
 const routes = require('./routes');
 const sequelize = require('./config/connection');
+const exphbs = require('express-handlebars');
+const helpers = require('./utils/helpers');
+
+const hbs = exphbs.create({ helpers });
 
 const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
