+++ views/homepage.handlebars
@@ -38,4 +38,4 @@
 </div>
 
 {{!-- Need to get the correct src filepath --}}
-<script src="./js/login.js"></script> 
+<script src="/js/login.js"></script>
