+++ views/layouts/main.handlebars
@@ -5,8 +5,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Exposition Inn</title>
-  <link rel="stylesheet" href="../../public/css/jass.css">
-  <link rel="stylesheet" href="../../public/css/style.css">
+  <link rel="stylesheet" href="/css/jass.css">
+  <link rel="stylesheet" href="/css/style.css">
 </head>
 
 <body>
@@ -21,13 +21,15 @@
         <a href="/profile">profile</a> | <!-- Need to add the correct pathing -->
         <button class="no-button" id="logout">logout</button>
         {{else}}
-        <a href="api/users/login">login</a>
+        <a href="/login">login</a>
         {{/if}}
       </nav>
     </header>
     <main class="container container-fluid mt-5">
       <!-- Render the sub layout -->
+
       {{{ body }}}
+
     </main>
     <footer>
       
@@ -36,7 +38,7 @@
 
   <!-- Render script for logged in users only -->
   {{#if logged_in}}
-  <script src="./js/logout.js"></script>
+  <script src="/js/logout.js"></script>
   {{/if}}
 </body>
