+++ views/user.handlebars
@@ -19,4 +19,4 @@
   {{/if}}
 </div>
 
-<script src="./js/user.js"></script>
+<script src="/js/user.js"></script>
